server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
pypinyin==0.53.0